pyproject.toml
[project]
name = "cs4all-backend"
version = "0.1.0"
description = "AI grading and progress-tracking backend for the cs4all Vietnamese CS learning platform."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # ── Web Framework ────────────────────────────────────────────────────────
    "fastapi[standard]>=0.115.0",

    # ── Settings & Validation ────────────────────────────────────────────────
    # pydantic-settings reads .env files and validates env vars at startup.
    # Prefer this over bare os.environ.get() — missing vars cause a clear error.
    "pydantic-settings>=2.4.0",

    # ── Structured Logging ───────────────────────────────────────────────────
    # JSON output in production, coloured console in development.
    "structlog>=24.4.0",

    # ── Database ─────────────────────────────────────────────────────────────
    "supabase>=2.28.0",

    # ── Task Queue ───────────────────────────────────────────────────────────
    # redis[hiredis] gives async support + the faster hiredis C parser.
    "redis[hiredis]>=5.0.0",

    # ── HTTP Client (Phase 3.5: GitHub API for rubric fetching) ─────────────
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    # ── Testing ────────────────────────────────────────────────────────────
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",          # AsyncClient for FastAPI TestClient
]
